Remove old logical schema
